Fix fatal error.
Accidentally moved a class method outside of the class in a context that
couldn't be called. This led to needing to define the constant URI as
the full URI. Additionally, defining a constant array is not possible,
so that needed to be changed as well. All of these updates are in order
to allow config values to override the defaults.