Update CSP for better plugin focus.
The CSP was flagging inline scripts and data URIs. These reports are
confusing as they do not produce mixed content warnings, but represent
another type of attack.

The CSP is changed to be more permissive for things that might represent
a XSS attack simply because that is not the aim of the plugin. By
allowing inline script and data URIs, the logs are cleaner and only
represent things that cause mixed content warnings.