Begin policy improvements.
This commit does two things:

1. The commit introduces `$mcd_policies`, which is a simpler to
understand array for setting the content policy. Reading a CSP can be a
little daunting as a long string. Breaking it into an array can make
things much simpler to comprehend and configure.

2. Allow the array to be overridden. The array will only use the default
if the user does not intentionally override it. This is paving the way
for an officially supported way to add one's own CSP.