Build CSP based on array of values.
A lot of work needed to be done to create the CSP string out of an array
of individual policies. This commit recognizes and implements the CSP
array as a string. Overall, this makes generating a policy dead simple.

Furthermore, the switch to individual policies will allow easier
categorization of violations so a user, for instance, can see all image
violations in a single grouping.