Decouple the report-uri from the policies.
The report-uri is not a policy and should not be treated as such. If a
user adds a custom policy, she should not need to specify the report-uri
as well. That should be handled for her. Additionally, I do not yet want
people adding custom report-uris as they just won't work without extra
effort on the part of the developer.