Add initial WP CLI command.
The first WP CLI command simply displays the list of violations much
like the WP List Table does. There is much to be done to make this look
a lot nicer and be more usable, but this is a start.

Note that the table will appear broke in WP CLI due to a bug in PHP CLI
Tools: https://github.com/wp-cli/php-cli-tools/pull/75.