Cleanup WP CLI table reporting.
The WP CLI table is really hard to read with the amount of data that is
produced. In order to help with this, the "Original Policy" column is
removed. It really does not provide that much information and it's so
much easier to read without it there. Since we show the "Violated
Directive", it's redundant to show the whole policy.

Additionally, "Report ID" was simplified to "ID", "Resolved" to "Status"
and "Valid HTTPS URI" to "HTTPS". This allows the header to sit on one
row when the terminal is roughly 100 columns wide, which is a reasonable
default.