Initial commit for sampling mode.
Sampling mode allows requests to be accepted outside of a logged in
request. Two new constants are added to turn sampling mode on/off, as
well as set the frequency of requests to sample.

Sampling mode works by setting the CSP header for all requests, but only
allowing a certain number of requests to be accepted by the beacon. The
frequency constant sets the number of requests that are accepted. This
method should help throttle the number of requests that get through.