Initial work on violation location detection.
This commit begins the articulation of an OOP pattern for defining
violation location checks. It articulates an interface that the checks
must conform to and a collector to house these checks. Much of this is
still being worked out. The initial violation location check is
scaffolded to check for the violation within post content.