Allow lookups for a single violation.
It would be convenient to use a simple API for gathering the necessary
data for a violation. Right now, the API only supports querying for all
reports. We need a way to get a single report.

This commit introduces querying by a single report ID. It'll be helpful
for times when only one report is needed.