Only allow matching a single violation location.
Violations that are located in raw content are most likely also going to
be found in filtered content; however, it is important to note that the
violation is in raw content, not filtered content as this gives a better
idea of where to resolve the violation. To fix this issue, we need to
exit the search once a violation is located. This also means that the
order of violation location registration is important.