Add content base class to reduce redundancy.
Since the two content search are identical other than their identity
information and the content that it searches, a base abstract class was
developed to share functionality with the individual classes. This
reduces the redundancy of the two classes and will make for easier
maintenance moving forward.

Some of these methods in the base class may prove to be helpful in the
violation base class. If so, further refactoring may be done.