Add the shortcode violation location.
The shortcode violation location attempts to find which shortcode caused
the blocked URI to be rendered. A good example of this is the "gallery"
shortcode. The ID of an image is saved as a shortcode attribute and the
post content does not contain a violation. Once the "gallery" shortcode
is applied and the URL is expressed, a violation is present. This
violation location will suss out these locations and specify the
shortcode that caused the issue.